+++ agent/storage/clone_test.go
@@ -1,0 +1,116 @@
+package storage
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"path/filepath"
+	"testing"
+
+	"github.com/erikmagkekse/btrfs-nfs-csi/agent/storage/nfs"
+	"github.com/erikmagkekse/btrfs-nfs-csi/config"
+	"github.com/erikmagkekse/btrfs-nfs-csi/utils"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+// --- TestCreateClone ---
+
+func TestCreateClone(t *testing.T) {
+	ctx := context.Background()
+
+	// setupSrcSnap creates a source snapshot with data/ dir and metadata.
+	setupSrcSnap := func(t *testing.T, bp, name string) {
+		t.Helper()
+		snapDir := filepath.Join(bp, config.SnapshotsDir, name)
+		dataDir := filepath.Join(snapDir, config.DataDir)
+		require.NoError(t, os.MkdirAll(dataDir, 0o755))
+		writeSnapshotMetadata(t, snapDir, SnapshotMetadata{Name: name, Volume: "srcvol", ReadOnly: true})
+	}
+
+	t.Run("validation", func(t *testing.T) {
+		tests := []struct {
+			name  string
+			req   CloneCreateRequest
+			setup bool
+			code  string
+		}{
+			{name: "invalid_name", req: CloneCreateRequest{Name: "bad!", Snapshot: "snap"}, code: ErrInvalid},
+			{name: "invalid_snapshot", req: CloneCreateRequest{Name: "clone", Snapshot: "bad!"}, code: ErrInvalid},
+			{name: "snapshot_not_found", req: CloneCreateRequest{Name: "clone", Snapshot: "nonexistent"}, code: ErrNotFound},
+			{name: "already_exists", req: CloneCreateRequest{Name: "existing", Snapshot: "mysnap"}, setup: true, code: ErrAlreadyExists},
+		}
+		for _, tt := range tests {
+			t.Run(tt.name, func(t *testing.T) {
+				s, bp, _, _ := newTestStorage(t)
+				if tt.setup || tt.req.Snapshot == "mysnap" {
+					setupSrcSnap(t, bp, "mysnap")
+				}
+				if tt.name == "already_exists" {
+					cloneDir := filepath.Join(bp, "existing")
+					require.NoError(t, os.MkdirAll(cloneDir, 0o755))
+					require.NoError(t, writeMetadataAtomic(
+						filepath.Join(cloneDir, config.MetadataFile),
+						CloneMetadata{Name: "existing", SourceSnapshot: "mysnap"},
+					))
+				}
+				meta, err := s.CreateClone(ctx, "test", tt.req)
+				requireStorageError(t, err, tt.code)
+				if tt.name == "already_exists" {
+					require.NotNil(t, meta, "should return existing metadata")
+					assert.Equal(t, "existing", meta.Name)
+				}
+			})
+		}
+	})
+
+	t.Run("success", func(t *testing.T) {
+		s, bp, runner, _ := newTestStorage(t)
+		setupSrcSnap(t, bp, "mysnap")
+
+		meta, err := s.CreateClone(ctx, "test", CloneCreateRequest{
+			Name: "myclone", Snapshot: "mysnap",
+		})
+		require.NoError(t, err, "CreateClone")
+		assert.Equal(t, "myclone", meta.Name)
+		assert.Equal(t, "mysnap", meta.SourceSnapshot)
+		assert.Equal(t, filepath.Join(bp, "myclone"), meta.Path)
+		assert.False(t, meta.CreatedAt.IsZero(), "CreatedAt should be set")
+
+		var ondisk CloneMetadata
+		require.NoError(t, ReadMetadata(filepath.Join(bp, "myclone", config.MetadataFile), &ondisk))
+		assert.Equal(t, "myclone", ondisk.Name, "on-disk metadata should match")
+
+		// btrfs snapshot called WITHOUT -r flag (writable clone)
+		srcData := filepath.Join(bp, config.SnapshotsDir, "mysnap", config.DataDir)
+		dstData := filepath.Join(bp, "myclone", config.DataDir)
+		require.Len(t, runner.Calls, 1, "expected exactly 1 btrfs call")
+		assert.Equal(t, []string{"subvolume", "snapshot", srcData, dstData}, runner.Calls[0])
+	})
+
+	t.Run("btrfs_fails_cleanup", func(t *testing.T) {
+		runner := &utils.MockRunner{Err: fmt.Errorf("snapshot error")}
+		exporter := &nfs.MockExporter{}
+		s, bp := testStorageWithRunner(t, runner, exporter)
+		setupSrcSnap(t, bp, "mysnap")
+
+		_, err := s.CreateClone(ctx, "test", CloneCreateRequest{
+			Name: "failclone", Snapshot: "mysnap",
+		})
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "btrfs snapshot failed")
+
+		cloneDir := filepath.Join(bp, "failclone")
+		_, statErr := os.Stat(cloneDir)
+		assert.True(t, os.IsNotExist(statErr), "cloneDir should be cleaned up after failure")
+	})
+
+	t.Run("invalid_tenant", func(t *testing.T) {
+		s, _, _, _ := newTestStorage(t)
+
+		_, err := s.CreateClone(ctx, "bad tenant!", CloneCreateRequest{
+			Name: "clone", Snapshot: "snap",
+		})
+		requireStorageError(t, err, ErrInvalid)
+	})
+}
